Calculator logic, added app icons

+++ CalculatorOperations/Sources/CalculatorOperations/CalculatorOperations.swift
@@ -3,7 +3,7 @@ import Foundation
 public protocol CalculatorOperation {
     var numberOfArguments: Int { get }
     var name: String { get }
-    func calculate(_ args:[CGFloat]) async throws -> CGFloat
+    func calculate(_ args:[Float]) async throws -> Float
 }
 
 public class AddOperation: CalculatorOperation {
@@ -12,7 +12,7 @@ public class AddOperation: CalculatorOperation {
     
     public init() {}
     
-    public func calculate(_ args: [CGFloat]) async -> CGFloat {
+    public func calculate(_ args: [Float]) async -> Float {
         return args.reduce(0) { partialResult, anotherArg in
              partialResult + anotherArg
         }
